Extracted url state sync code into a hook
